fix: correctness and stability fixes

+++ src/SubnetSearch.Core/Models/Classification/IpPrefix.cs
@@ -4,5 +4,7 @@ public record IpPrefix(
     string  Prefix,
     string? CountryCode,
     string? Description,
-    int     IpCount
+    // Host count of the prefix. A /1 covers 2^31 and a /0 covers 2^32 addresses,
+    // neither of which fits in a 32-bit int (F20) — use long.
+    long    IpCount
 );
